Remove Manager class in main
