+++ src/main/java/seedu/address/model/util/SampleDataUtil.java
@@ -21,7 +21,6 @@ import seedu.address.model.person.Username;
 import seedu.address.model.route.ReadOnlyRouteList;
 import seedu.address.model.route.Route;
 import seedu.address.model.route.RouteList;
-import seedu.address.model.user.Manager;
 import seedu.address.model.user.User;
 
 /**
@@ -90,11 +89,11 @@ public class SampleDataUtil {
     }
 
     public static User[] getSampleUsers() {
-        return new User[]{
-            new Manager(new Name("Alice Pauline"), new Username("alicepauline"), new Password("alicepauline01")),
-            new Manager(new Name("Benson Meier"), new Username("bensonmeier"), new Password("bensonmeier02")),
-            new Manager(new Name("Carl Kurz"), new Username("carlkurz"), new Password("carlkurz03")),
-            new Manager(new Name("Hoon Meier"), new Username("hoonmeier"), new Password("hoonmeier04"))
+        return new User[] {
+            new User(new Name("Alice Pauline"), new Username("alicepauline"), new Password("alicepauline01")),
+            new User(new Name("Benson Meier"), new Username("bensonmeier"), new Password("bensonmeier02")),
+            new User(new Name("Carl Kurz"), new Username("carlkurz"), new Password("carlkurz03")),
+            new User(new Name("Hoon Meier"), new Username("hoonmeier"), new Password("hoonmeier04"))
         };
     }
